It just occured to me
I’ve got a bunch of private keys for video apps that are checked into
version control.

That’s bad.

Anyway, a new endpoint as been added: /invalidate-vimeo-cache, which
can be used to manually invalidate vimeo if we’ve made a change and we
don’t want to wait six hours for it to show on the site.

This will be removed once the website gets its data from gears.

Also the signature is now served in /assets/gm-2017-signature.png