Freewall alternative
I’ve devised my own Freewall alternative.

It’s much faster than the jQuery predecessor.

The code is a bit messy, it still has some things to fix, and it
doesn’t have nearly as many options as jQuery freewall, but it still
works.

Was easier than I thought it would be, too.